requirements.txt
# 경제 데이터 모니터링 시스템 의존성

# 데이터 수집 및 분석
yfinance>=0.2.18
pandas>=1.5.0
numpy>=1.24.0

# 비동기 처리
aiohttp>=3.8.0
asyncio

# AWS 서비스
boto3>=1.26.0
botocore>=1.29.0

# Agent 시스템 및 LLM
langchain>=0.2.11
langchain-aws>=0.1.16
langchain-community>=0.2.10
pydantic>=2.8.2
typing-extensions>=4.12.2

# 웹 스크래핑 및 데이터 처리
requests>=2.28.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
selenium>=4.20.0
webdriver-manager>=4.0.0

# 로깅 및 유틸리티
python-dateutil>=2.8.0
pytz>=2023.3
python-dotenv>=1.0.0
schedule>=1.2.0

# Streamlit 및 시각화
streamlit>=1.28.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
altair>=5.0.0

# 이미지 생성 및 처리
Pillow>=10.0.0
opencv-python>=4.8.0
wordcloud>=1.9.0

# 시스템 모니터링을 위한 추가 의존성
psutil>=5.9.0

# Slack 알림 시스템을 위한 추가 의존성
aiohttp>=3.8.0

# 고도화된 분석을 위한 추가 의존성
scipy>=1.10.0
scikit-learn>=1.3.0
feedparser>=6.0.10

# 기존 프로젝트 의존성 (이미 설치된 것들)
# torch
# networkx
# matplotlib
# plotly
# bokeh

# 선택적 의존성 (추후 확장용)
# alpha-vantage  # Alpha Vantage API
# fredapi       # FRED API
